Various small fixes followning the more pedantic gcc flags (see https://github.com/airbus-seclab/c-compiler-security for a good reference on these flags):
-Wall -Wextra -Wpedantic -Wformat=2 -Wformat-overflow=2 -Wformat-truncation=2
-Wformat-security -Wnull-dereference -Wstack-protector -Wtrampolines -Walloca
-Wvla -Warray-bounds=2 -Wimplicit-fallthrough=3 -Wtraditional-conversion
-Wshift-overflow=2 -Wcast-qual -Wstringop-overflow=4 -Wconversion
-Warith-conversion -Wlogical-op -Wduplicated-cond -Wduplicated-branches
-Wformat-signedness -Wshadow -Wstrict-overflow=2 -Wundef -Wstrict-prototypes
-Wswitch-default -Wstack-usage=1000000 -Wcast-align=strict

NOTE: -Wstrict-overflow >= 3 gives false positives, and -Wswitch-enum is
almost unusable for some cases in libecc code (explicitely enumerating all the
enum values make the code less readable). This is why these two flags are not
used.

+++ src/curves/aff_pt.c
@@ -243,7 +243,7 @@ int aff_pt_import_from_buf(aff_pt_t pt,
 	ret = ec_shortw_crv_check_initialized(crv); EG(ret, err);
 
 	ctx = crv->a.ctx;
-	coord_len = BYTECEIL(ctx->p_bitlen);
+	coord_len = (u16)BYTECEIL(ctx->p_bitlen);
 
 	MUST_HAVE((pt_buf_len == (2 * coord_len)), ret, err);
 
@@ -295,7 +295,7 @@ int aff_pt_export_to_buf(aff_pt_src_t pt, u8 *pt_buf, u32 pt_buf_len)
 	MUST_HAVE((on_curve), ret, err);
 
 	/* buffer size must match 2 * p_len */
-	coord_len = BYTECEIL(pt->crv->a.ctx->p_bitlen);
+	coord_len = (u16)BYTECEIL(pt->crv->a.ctx->p_bitlen);
 	MUST_HAVE((pt_buf_len == (2 * coord_len)), ret, err);
 
 	/* Export the two coordinates */
