+++ src/curves/ec_params.c
@@ -98,7 +98,7 @@ int import_params(ec_params *out_params, const ec_str_params *in_str_params)
 	ret = nn_init_from_buf(&tmp_order,
 			 PARAM_BUF_PTR(in_str_params->gen_order),
 			 PARAM_BUF_LEN(in_str_params->gen_order)); EG(ret, err);
-	ret = nn_init(&(out_params->ec_gen_order), tmp_order.wlen * WORD_BYTES); EG(ret, err);
+	ret = nn_init(&(out_params->ec_gen_order), (u16)(tmp_order.wlen * WORD_BYTES)); EG(ret, err);
 	ret = nn_copy(&(out_params->ec_gen_order), &tmp_order); EG(ret, err);
 
 	ret = nn_init_from_buf(&tmp_order_bitlen,
@@ -110,7 +110,7 @@ int import_params(ec_params *out_params, const ec_str_params *in_str_params)
 			 PARAM_BUF_PTR(in_str_params->cofactor),
 			 PARAM_BUF_LEN(in_str_params->cofactor)); EG(ret, err);
 	ret = nn_init(&(out_params->ec_gen_cofactor),
-		tmp_cofactor.wlen * WORD_BYTES); EG(ret, err);
+		(u16)(tmp_cofactor.wlen * WORD_BYTES)); EG(ret, err);
 	ret = nn_copy(&(out_params->ec_gen_cofactor), &tmp_cofactor); EG(ret, err);
 
 	/* Now we can store the number of points on the curve */
