+++ src/fp/fp.c
@@ -152,7 +152,7 @@ int fp_init(fp_t in, fp_ctx_src_t fpctx)
 	MUST_HAVE((in != NULL), ret, err);
 
 	ret = fp_ctx_check_initialized(fpctx); EG(ret, err);
-	ret = nn_init(&(in->fp_val), (fpctx->p.wlen) * WORD_BYTES); EG(ret, err);
+	ret = nn_init(&(in->fp_val), (u16)((fpctx->p.wlen) * WORD_BYTES)); EG(ret, err);
 
 	in->ctx = fpctx;
 	in->magic = FP_MAGIC;
