+++ src/sig/ec_key.h
@@ -140,7 +140,7 @@ typedef struct {
 #error "All structured pub keys size are expected to fit on an u8."
 #endif
 #define EC_STRUCTURED_PUB_KEY_EXPORT_SIZE(pub_key)			\
-	((u8)(EC_PUB_KEY_EXPORT_SIZE(pub_key) + (3 * sizeof(u8))))
+	((u8)(EC_PUB_KEY_EXPORT_SIZE(pub_key) + (u8)(3 * sizeof(u8))))
 
 ATTRIBUTE_WARN_UNUSED_RET int pub_key_check_initialized(const ec_pub_key *A);
 ATTRIBUTE_WARN_UNUSED_RET int pub_key_check_initialized_and_type(const ec_pub_key *A,
